refactor: split AT-token template helpers out of executor_lastcallers.go
